finished initial impl of reading from cat/dat files

+++ trunk/MFC Test 1/GZipStream.h
@@ -46,7 +46,7 @@ namespace Library
       public:
          void   Close();
          void   SafeClose();
-         void   Seek(DWORD  offset, SeekOrigin  mode);
+         void   Seek(LONG  offset, SeekOrigin  mode);
          void   SetLength(DWORD  length);
 
          DWORD  Read(BYTE* buffer, DWORD length);
